ai-agent-support-system/requirements.txt
fastapi
uvicorn
python-dotenv

langchain
langgraph
langchain-community
langchain-google-genai
langchain-text-splitters

faiss-cpu
sentence-transformers
huggingface-hub
pydantic
